refactor: change the initialization logic
